+++ frontend/src/components/slc-configurate-function/components/components.ts
@@ -1,0 +1,2 @@
+export { CreateForm } from './create-form/create-form';
+export { EditForm } from './edit-form/edit-form';
