+++ frontend/src/components/slc-configurate-function/components/edit-form/styles.module.scss
@@ -1,0 +1,15 @@
+.formTitle {
+  margin: 0 0 15px;
+  color: var(--color-text);
+  font-size: 30px;
+}
+
+.buttons {
+  display: flex;
+  justify-content: flex-end;
+  gap: 10px;
+}
+
+.button {
+  min-width: 200px;
+}
